wip: refactored and added currentPhase on Game
